Move internal/safepath to daemon/volume/safepath
