+++ daemon/volume/mounts/mounts.go
@@ -9,7 +9,7 @@ import (
 	"github.com/containerd/log"
 	mounttypes "github.com/docker/docker/api/types/mount"
 	"github.com/docker/docker/daemon/volume"
-	"github.com/docker/docker/internal/safepath"
+	"github.com/docker/docker/daemon/volume/safepath"
 	"github.com/docker/docker/pkg/idtools"
 	"github.com/docker/docker/pkg/stringid"
 	"github.com/moby/sys/user"
