+++ daemon/volume/safepath/common.go
