+++ daemon/volume/safepath/errors.go
