+++ daemon/volume/safepath/join_linux.go
