+++ daemon/volume/safepath/safepath.go
