+++ integration/volume/mount_test.go
@@ -17,8 +17,8 @@ import (
 	"github.com/docker/docker/api/types/versions"
 	"github.com/docker/docker/api/types/volume"
 	"github.com/docker/docker/client"
+	"github.com/docker/docker/daemon/volume/safepath"
 	"github.com/docker/docker/integration/internal/container"
-	"github.com/docker/docker/internal/safepath"
 	"github.com/docker/docker/testutil/fakecontext"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
